tools/requirements-local-ai.txt
flask>=3.0.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
Pillow>=10.0.0
paddleocr>=2.7.0
faster-whisper>=1.1.0
